myhomepage::article - a class to handle articles
The directory public/articles is the root for all articles.
Each article is a directory, whose name is the slug of the article.

If the article is hosted in the public/articles top-directory, it is
considered a page.

If the article is under another level of directory (like
public/articles/tech/some-slug), then its parent directory is considered
to be its category.

Hence an article belongs to at most one category.

Within the article directory, some files are expected to be found:
  * meta.yml : contains the meta data of the article such as date,
    author, tags, etc.
  * contend.md : the MarkDown content of the article itself.