Add the blog engine - articles and permalink
The homepage displays the "featured" articles listed in the
blog-meta.yml file, hosted at the root of the blog: public/articles

If none found, the section is not displayed.

Each article found is displayed with a blog-card on the home and links
to its permalink page.

The content of the blog post itself is to be hosted in a dedicated
directory whose name is the slug.
  * content is looked for in <slug>/contend.md
  * meta data is looked for in <slug>/meta.yml